feat: remove all original animation data structures
